+++ mobile/android/android-components/components/feature/top-sites/src/main/java/mozilla/components/feature/top/sites/db/PinnedSiteEntity.kt
@@ -8,8 +8,6 @@ import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import mozilla.components.feature.top.sites.TopSite
-import mozilla.components.feature.top.sites.TopSite.Type.DEFAULT
-import mozilla.components.feature.top.sites.TopSite.Type.PINNED
 
 /**
  * Internal entity representing a pinned site.
@@ -32,16 +30,22 @@ internal data class PinnedSiteEntity(
     @ColumnInfo(name = "created_at")
     var createdAt: Long = System.currentTimeMillis()
 ) {
-    internal fun toTopSite(): TopSite {
-        val type = if (isDefault) DEFAULT else PINNED
-        return TopSite(
-            id,
-            title,
-            url,
-            createdAt,
-            type
-        )
-    }
+    internal fun toTopSite(): TopSite =
+        if (isDefault) {
+            TopSite.Default(
+                id = id,
+                title = title,
+                url = url,
+                createdAt = createdAt
+            )
+        } else {
+            TopSite.Pinned(
+                id = id,
+                title = title,
+                url = url,
+                createdAt = createdAt
+            )
+        }
 }
 
 internal fun TopSite.toPinnedSite(): PinnedSiteEntity {
@@ -49,7 +53,7 @@ internal fun TopSite.toPinnedSite(): PinnedSiteEntity {
         id = id,
         title = title ?: "",
         url = url,
-        isDefault = type === DEFAULT,
+        isDefault = this is TopSite.Default,
         createdAt = createdAt ?: 0
     )
 }