+++ mobile/android/android-components/components/feature/top-sites/src/main/java/mozilla/components/feature/top/sites/ext/TopFrecentSiteInfo.kt
@@ -6,18 +6,16 @@ package mozilla.components.feature.top.sites.ext
 
 import mozilla.components.concept.storage.TopFrecentSiteInfo
 import mozilla.components.feature.top.sites.TopSite
-import mozilla.components.feature.top.sites.TopSite.Type.FRECENT
 import mozilla.components.support.ktx.kotlin.tryGetHostFromUrl
 
 /**
  * Returns a [TopSite] for the given [TopFrecentSiteInfo].
  */
 fun TopFrecentSiteInfo.toTopSite(): TopSite {
-    return TopSite(
+    return TopSite.Frecent(
         id = null,
         title = this.title?.takeIf(String::isNotBlank) ?: this.url.tryGetHostFromUrl(),
         url = this.url,
-        createdAt = null,
-        type = FRECENT
+        createdAt = null
     )
 }