+++ mobile/android/android-components/components/feature/top-sites/src/test/java/mozilla/components/feature/top/sites/DefaultTopSitesStorageTest.kt
@@ -72,35 +72,32 @@ class DefaultTopSitesStorageTest {
             coroutineContext = coroutineContext
         )
 
-        val frecentSite = TopSite(
+        val frecentSite = TopSite.Frecent(
             id = 1,
             title = "Mozilla",
             url = "https://mozilla.com",
-            createdAt = 1,
-            type = TopSite.Type.FRECENT
+            createdAt = 1
         )
         defaultTopSitesStorage.removeTopSite(frecentSite)
 
         verify(historyStorage).deleteVisitsFor(frecentSite.url)
 
-        val pinnedSite = TopSite(
+        val pinnedSite = TopSite.Pinned(
             id = 2,
             title = "Firefox",
             url = "https://firefox.com",
-            createdAt = 2,
-            type = TopSite.Type.PINNED
+            createdAt = 2
         )
         defaultTopSitesStorage.removeTopSite(pinnedSite)
 
         verify(pinnedSitesStorage).removePinnedSite(pinnedSite)
         verify(historyStorage).deleteVisitsFor(pinnedSite.url)
 
-        val defaultSite = TopSite(
+        val defaultSite = TopSite.Default(
             id = 3,
             title = "Wikipedia",
             url = "https://wikipedia.com",
-            createdAt = 3,
-            type = TopSite.Type.DEFAULT
+            createdAt = 3
         )
         defaultTopSitesStorage.removeTopSite(defaultSite)
 
@@ -117,34 +114,31 @@ class DefaultTopSitesStorageTest {
             coroutineContext = coroutineContext
         )
 
-        val defaultSite = TopSite(
+        val defaultSite = TopSite.Default(
             id = 1,
             title = "Firefox",
             url = "https://firefox.com",
-            createdAt = 1,
-            type = TopSite.Type.DEFAULT
+            createdAt = 1
         )
         defaultTopSitesStorage.updateTopSite(defaultSite, "Mozilla Firefox", "https://mozilla.com")
 
         verify(pinnedSitesStorage).updatePinnedSite(defaultSite, "Mozilla Firefox", "https://mozilla.com")
 
-        val pinnedSite = TopSite(
+        val pinnedSite = TopSite.Pinned(
             id = 2,
             title = "Wikipedia",
             url = "https://wikipedia.com",
-            createdAt = 2,
-            type = TopSite.Type.PINNED
+            createdAt = 2
         )
         defaultTopSitesStorage.updateTopSite(pinnedSite, "Wiki", "https://en.wikipedia.org/wiki/Wiki")
 
         verify(pinnedSitesStorage).updatePinnedSite(pinnedSite, "Wiki", "https://en.wikipedia.org/wiki/Wiki")
 
-        val frecentSite = TopSite(
+        val frecentSite = TopSite.Frecent(
             id = 1,
             title = "Mozilla",
             url = "https://mozilla.com",
-            createdAt = 1,
-            type = TopSite.Type.FRECENT
+            createdAt = 1
         )
         defaultTopSitesStorage.updateTopSite(frecentSite, "Moz", "")
 
@@ -160,19 +154,17 @@ class DefaultTopSitesStorageTest {
             coroutineContext = coroutineContext
         )
 
-        val defaultSite = TopSite(
+        val defaultSite = TopSite.Default(
             id = 1,
             title = "Firefox",
             url = "https://firefox.com",
-            createdAt = 1,
-            type = TopSite.Type.DEFAULT
+            createdAt = 1
         )
-        val pinnedSite = TopSite(
+        val pinnedSite = TopSite.Pinned(
             id = 2,
             title = "Wikipedia",
             url = "https://wikipedia.com",
-            createdAt = 2,
-            type = TopSite.Type.PINNED
+            createdAt = 2
         )
         whenever(pinnedSitesStorage.getPinnedSites()).thenReturn(
             listOf(
@@ -213,19 +205,17 @@ class DefaultTopSitesStorageTest {
             coroutineContext = coroutineContext
         )
 
-        val defaultSite = TopSite(
+        val defaultSite = TopSite.Default(
             id = 1,
             title = "Firefox",
             url = "https://firefox.com",
-            createdAt = 1,
-            type = TopSite.Type.DEFAULT
+            createdAt = 1
         )
-        val pinnedSite = TopSite(
+        val pinnedSite = TopSite.Pinned(
             id = 2,
             title = "Wikipedia",
             url = "https://wikipedia.com",
-            createdAt = 2,
-            type = TopSite.Type.PINNED
+            createdAt = 2
         )
         whenever(pinnedSitesStorage.getPinnedSites()).thenReturn(
             listOf(
@@ -307,26 +297,23 @@ class DefaultTopSitesStorageTest {
             coroutineContext = coroutineContext
         )
 
-        val defaultSiteFirefox = TopSite(
+        val defaultSiteFirefox = TopSite.Default(
             id = 1,
             title = "Firefox",
             url = "https://firefox.com",
-            createdAt = 1,
-            type = TopSite.Type.DEFAULT
+            createdAt = 1
         )
-        val pinnedSite1 = TopSite(
+        val pinnedSite1 = TopSite.Pinned(
             id = 2,
             title = "Wikipedia",
             url = "https://wikipedia.com",
-            createdAt = 2,
-            type = TopSite.Type.PINNED
+            createdAt = 2
         )
-        val pinnedSite2 = TopSite(
+        val pinnedSite2 = TopSite.Pinned(
             id = 3,
             title = "Example",
             url = "https://example.com",
-            createdAt = 3,
-            type = TopSite.Type.PINNED
+            createdAt = 3
         )
         whenever(pinnedSitesStorage.getPinnedSites()).thenReturn(
             listOf(