+++ mobile/android/android-components/docs/changelog.md
@@ -15,6 +15,7 @@ permalink: /changelog/
   * ⚠️ **This is a breaking change**: Removed `String.urlToTrimmedHost` extension method.
 
 * **feature-top-sites**
+  * ⚠️ **This is a breaking change**: The existing data class `TopSite` has been converted into a sealed class. [#11483](https://github.com/mozilla-mobile/android-components/issues/11483)
   * Extend `DefaultTopSitesStorage` to accept a `TopSitesProvider` for fetching top sites. [#11483](https://github.com/mozilla-mobile/android-components/issues/11483)
 
 # 97.0.0